CI: Stop auto-releases until next Cascadia has been evaluated
[why]
The next Caskadia Code release will (probably) bring a native Nerd Fonts
version with it. We should examine and evaluate it and the new subfonts
before we release a new Delugia.

Also Delugia run still on a rather old Nerd Fonts version, so if we want
to release a new Delugia we should probably update the Nerd Font part
first.

We can reinstate this if we have a plan what to do and why.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>